feat: added Layout component + shopItem route
