+++ package.json
@@ -25,6 +25,7 @@
     "resolve-url-loader": "^5.0.0",
     "sass": "^1.69.7",
     "slick-carousel": "^1.8.1",
+    "uuid": "^9.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
