+++ src/components/layout/Layout.jsx
@@ -1,0 +1,14 @@
+import Header from "../Header";
+import Footer from "../Footer";
+
+const Layout = ({ children }) => {
+  return (
+    <main>
+      <Header />
+      {children}
+      <Footer />
+    </main>
+  );
+};
+
+export default Layout;
