+++ src/pages/shop/data.js
@@ -1,52 +1,75 @@
+import { v4 as uuidv4 } from "uuid";
+
 const mockwines = [
   {
+    id: uuidv4(),
     name: "Cabernet Sauvignon",
     price: 30.99,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "California, USA",
+    description:
+      "A rich and full-bodied red wine with notes of blackcurrant and cedar.",
   },
   {
+    id: uuidv4(),
     name: "Chardonnay",
     price: 25.49,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Burgundy, France",
+    description:
+      "A crisp and elegant white wine with flavors of green apple and citrus.",
   },
   {
+    id: uuidv4(),
     name: "Merlot",
     price: 22.99,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Tuscany, Italy",
+    description:
+      "A smooth and velvety red wine with hints of plum and chocolate.",
   },
   {
+    id: uuidv4(),
     name: "Pinot Noir",
     price: 35.75,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Oregon, USA",
+    description:
+      "An elegant and delicate red wine with flavors of cherry and earth.",
   },
   {
+    id: uuidv4(),
     name: "Sauvignon Blanc",
     price: 27.5,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Marlborough, New Zealand",
+    description:
+      "A vibrant and refreshing white wine with tropical fruit and herbal notes.",
   },
   {
+    id: uuidv4(),
     name: "Pinot Noir",
     price: 35.75,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Oregon, USA",
+    description:
+      "An elegant and delicate red wine with flavors of cherry and earth.",
   },
   {
+    id: uuidv4(),
     name: "Sauvignon Blanc",
     price: 27.5,
     imageLink:
       "https://marketwines.ca/wp-content/uploads/2023/02/Quintorelli-Riserva-2003.jpg",
     location: "Marlborough, New Zealand",
+    description:
+      "A vibrant and refreshing white wine with tropical fruit and herbal notes.",
   },
 ];
